requirements.txt
pyspark==3.5.5
pathling==7.2.0
delta-spark==3.3.0
typed-settings==24.6.0
typed-settings[attrs]==24.6.0
loguru==0.7.3
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-prometheus==0.50b0
